requirements.txt
docopt~=0.6.2
pyserial~=3.4
boto3~=1.16.19
pytest
freezegun
tox
rich